Moving numbers to internal subdirectory.
