+++ internal/numbers/trig/trig_test.go
