+++ widgets/linechart/zoom/zoom.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/mum4k/termdash/internal/mouse"
 	"github.com/mum4k/termdash/internal/mouse/button"
-	"github.com/mum4k/termdash/numbers"
+	"github.com/mum4k/termdash/internal/numbers"
 	"github.com/mum4k/termdash/terminalapi"
 	"github.com/mum4k/termdash/widgets/linechart/axes"
 )
