+++ widgets/sparkline/sparks.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 
 	"github.com/mum4k/termdash/internal/cell/runewidth"
-	"github.com/mum4k/termdash/numbers"
+	"github.com/mum4k/termdash/internal/numbers"
 )
 
 // sparks are the characters used to draw the SparkLine.
